[dace] Fix test_different_vertical_sizes

+++ tests/next_tests/integration_tests/multi_feature_tests/iterator_tests/test_column_stencil.py
@@ -294,7 +294,7 @@ def test_different_vertical_sizes(program_processor):
     k_size = 10
     inp0 = gtx.np_as_located_field(KDim)(np.asarray(list(range(k_size))))
     inp1 = gtx.np_as_located_field(KDim)(np.asarray(list(range(k_size + 1))))
-    out = gtx.np_as_located_field(KDim)(np.zeros(k_size))
+    out = gtx.np_as_located_field(KDim)(np.zeros(k_size, dtype=int))
     ref = inp0 + inp1[1:]
 
     run_processor(
@@ -334,7 +334,7 @@ def test_different_vertical_sizes_with_origin(program_processor):
     k_size = 10
     inp0 = gtx.np_as_located_field(KDim)(np.asarray(list(range(k_size))))
     inp1 = gtx.np_as_located_field(KDim, origin={KDim: 1})(np.asarray(list(range(k_size + 1))))
-    out = gtx.np_as_located_field(KDim)(np.zeros(k_size))
+    out = gtx.np_as_located_field(KDim)(np.zeros(k_size, dtype=int))
     ref = inp0 + np.asarray(inp1)[:-1]
 
     run_processor(
